Include indicator styling in the interface feature
For the large majority of skins, especially those based on the
MonoBook layout, page indicators are located in the h1, floated
to the right. Add the shared styles for this to the interface feature,
which is documented as
"Essentially this level is for styles that are common to MonoBook clones."
making it a suitable place to contain these styles.

Bug: T199035
Change-Id: Ib2087fc722da137f3e2677ac68a24333e33fb45d