Use localized interpunction for mediawiki.hlist
This converts it to Less, to access the messages.
Old IE8 style rules have been removed, which allows
simplifying the many style rules into a Less stylesheet
that doesn't repeat itself everywhere.

There is no message for " .", so that one has been marked
with a @todo, like in mediawiki.interface.helpers.styles.less.

Bug: T68336
Bug: T258766
Change-Id: I53bd9c55b4475113894b15d65be0ec02f1177770