mediawiki.base: Catch exceptions between mw.hook() handlers
Exceptions from one mw.hook handler should not affect another.
This expectation stems from how other systems work, such as
jQuery.ready, Promise, and DOM event listeners.

mw.hook used to be a shortcut to $.Callbacks(), which is the same API
that powers jQuery.ready and $.Deferred, but it is lower level and
before such error handling is used. With commit 9f23dfa47b4 I have
replaced with with a simple handling of our own with just the part
we need, and this commit improved that implementation with error
handling.

Bug: T223352
Change-Id: I380b93ab3de1b3b7e5f0c5effa2e367fa9b1e6f3