Move several legacy style rules to existing dedicated modules
These are often two or three css rules and can be included in the
style sheets of other modules.

To facilitate a good migration from mw-warning-with-logexcerpt,
mw-logevent-loglines has been added to the log event list of
LogEventsList, to allow styling previously provided under
mw-warning-with-logexcerpt to be selected only in cases where
the styles should actually be applied. See also T278002.

Bug: T278576
Change-Id: I3f45c95d2856c8fce2da5175b52089d4e629a41e