Drop DatabaseUpdater::doActiveUsersInit
This looks useful at the first glance but it's actually not going to do
anything unless if you're upgrading from a really old version.

The method checks and if active users is empty, it tries to populate it.
This can be the case when the upgrade is happening from a really old
version which we don't support direct upgrade anymore. Otherwise, with
each edit, the table gets updated and this won't be ran (unless the wiki
is completely inactive which is an edge this patch is going to fix).

Bug: T272199
Change-Id: Ia0bcc068a2f334e03824b3c0511734e0964ef092