Actor migration cleanup for removeUnusedAccounts.php
The query does a LEFT JOIN between user and actor. Given the state of
the actor table migration, it's impossible for a user to be active
without having an actor row. So such users must be inactive.

isInactiveAccount() does a loop over tables, sending them all to
ActorMigration, but that's obsolete for all except revision. So split
out the revision case out of the loop, with roughly the same code as
before. Inside the loop, do new-style queries in the knowledge that the
*_actor fields exist.

logging still needs a special case, but like the other non-revision
tables, it no longer needs ActorMigration.

Bug: T278917
Change-Id: I37281a0c856618674160bb39506108d5fff92479