Clean up and simplify SpecialEmailUser::submit
In preparation for moving to the EmailUserUtils service
being added in [1], but can be done in parallel.

Includes using some clearer variable names, and
using the Emailer "service" (which is just a wrapper
for UserMailer::send, but one that can be injected
and thus something that will be used when the
code is moved to the EmailUserUtils service,
changing the code now to make it easier to follow)

Additionally, expand Emailer::send to accept
a single MailAddress as the to, like
UserMailer::send.

Includes removing support for onEmailUser hook
setting $error to something other than a
Status object, true, false, an empty string, an array,
or a MessageSpecifier, which has been deprecated
since 1.29

[1] Ie63ddcc9805f4dbaa7a33e780611d9de227b0dc1

Bug: T265541
Change-Id: Iad683060f739f4fd5ebfd56d19702ba0129f8538