changes: Move unlock to cover DELETE query for activeusers in RC jobs
Follows-up bf8c90538dff567e.

The RecentChangesUpdateJob class has two modes of operation, one for
purging RC rows, and one for updating activeusers.

The former did its delete query within the lock, the latter did not.
We think this is the reason the latter sometimes causes a deadlock.
Move it within the lock to attempt to mitigate this.

Bug: T255059
Change-Id: Idc076dca15f4ba2e0a3a8f89623cfbfc9c25b27e