Bring back the CollationUpdate
Calls to doCollationUpdate was removed as it was introduced pre 1.27 and
not needed anymore which is incorrect, updating collation is important
work and it's needed every time you update the ICU version.

The underlying problem was that it was added as a ordinary update, while
it should be in a central place alongside other central updates like
checkStats, etc. So Instead of bringing them back to MySQLUpdater, this
change treat it as the same as other central updates.

Since we are here, removing the unneeded check that if the field exists,
this field was introduced pre 1.27 so no need to check its existence.

Bug: T272199
Change-Id: I5d809e525cfccf358a6545f2097549e942b8af0f