exception: Restore "PHP Deprecated" prefix for E_USER_DEPRECATED
Follows-up f2543d442a019, in which I accidentally removed the prefix.

Also, while at it, drop the "$class: " prefix that we normally
add to exceptions. For errors we internally construct ErrorException,
but this is something a clas name rarely seen by PHP developers and
has sometimes led to confusions about whether this is an exception
or not. We did still have "PHP Notice" and "PHP Warning" right after
it (except for deprecations) which should act as confidence indicator
that they are in fact not exceptions. We also have the channel
name ("error" instead of "exception"). But removing this will help
boost that confidence further, plus it had no added value anyhow, so
less is more?

Change-Id: Ib93f48c94d642f519558aab40d143b43e1d9ed16