Make SpecialFileDuplicateSearch not be a QueryPage
Since 2011, FileDuplicateSearch has been especially awkward as a
QueryPage subclass. The only thing of value in the parent was
construction of <ol></ol> tags, but that's not enough to justify it.
getQueryInfo() was unused  -- my motivation is to not have to maintain
it. A LoadBalancer was injected but was unused.

Change-Id: I2f0c18913260ee74ef266352bb93cfbdc1143d85