selenium: DEBUG environment variable increases Mocha timeout
By default, Mocha will abort the test suite if it takes more than 60
seconds. That is usually plenty when you're just running tests, but it's
usually not enough when you're debugging. At the moment, you have to
manually increase `mochaOpts` `timeout` in `wdio.conf.js`.

This patch checks if DEBUG is set, and if so, increases `timeout` from
60 seconds to 60 minutes.

This patch simplifies debugging. Instead of having to manually edit
`wdio.conf.js`, increasing `timeout` is done by:

    DEBUG=true npm run selenium-test

Debugging is documented at
https://www.mediawiki.org/wiki/Selenium/How-to/Debug_with_browser.debug()#mochaOpts

Bug: T278827
Change-Id: Ied8e7eab34d9472b841b803ada6541dd12f85976