ResourceLoaderSkinModule: When `toc` is enabled messages are empty
Only the `toc` feature requires access to messages. When that feature
is not enabled, do not load unnecessary messages.

Note, this will not fix the installer right away,
but will be needed to fix the
installer in addition to the change in
I0e3ba3c2cfa64bba49e3944d93e0c802b553bac8.

Bug: T270027
Change-Id: I3905bc2c3244ae2308c641bbabb097f62c666041