Remove MovePageFactoryTest
The tests in this file make no sense now that MovePageFactory no longer
exists. Obviously you do not pass all constructor arguments to
PageCommandFactory through to MovePage.

To make it pass against the following commit, I could update the
getExtraClassArgCount(), but that would be necessary pretty much every
time the constructor argument list of either class is changed. There's
no reason to think a difference in argument count is an error anymore.

Change-Id: Id4f4b43d47217ebb0111800d7ccb9855d56f7abb