Setup: Move LBFactory::setRequestInfo() to service wiring
Avoid costly and unconditional set up of LBFactory early on in
Setup.php. Instead, inject this information lazily if and when the
service is contructed by something that actually needs a database
connection.

This removes the cost of LBFactory service + dependencies for requests
to load.php, api.php, and rest.php requests that involve no (other)
calls to the database service.

Bug: T228895
Change-Id: I3140922cec130cad84eb89b2b34520c0988027dc