ImagePage: Add 2048px to $wgImageLimits for file description pages
1024px is becoming small a world of HiDPI screens. Wikimedia Commons
has worked around for this for over ten years with a JavaScript gadget
presenting links to 2000px. Let's standardise a similar size here so
as to obsolete that gadget.

Note that ThumbnailRenderJob (which is often a performance concern,
since it pre-renders thumbnails for all new uploads), is based on
$wgThumbLimits, not $wgImageLimits.

Bug: T106263
Change-Id: If3be7edc260c69aaedbe87e3ce8f578af57dbda6
(cherry picked from commit bcfb945cf43224524238db35d7a672bf02b203cf)