ForkController: Throw more meaningful error for missing PHP extensions
Currently all the signal constants emit warnings as they are undefined
in PHP 7.4 and throw fatal error in PHP 8, if pcntl is not available.

pcntl extension is not enabled by default, and is not required by
MediaWiki, so this class should throw a better error if it's not
found. The class already has comment that pcntl and posix are required
but this is meaningless since the error and a warning for each constant
would be emitted anyway.

https://www.php.net/manual/en/pcntl.installation.php

Convert RESTARTABLE_SIGNALS constant back to static property as it was,
because we can only use these constants after confirming they exist,
which happens in the constructor.

Bug: T280456
Change-Id: I66deaa9b346b936b2628ac0511492d20a42fea6c