Reorder the letters in the Saraiki alphabet linkTrail
Reordering according to the alphabet as it appears in the book
Bashir, Elena; Conners, Thomas J.; Hefright, Brook (2019).
A descriptive grammar of Hindko, Panjabi, and Saraiki.
De Gruyter Mouton. pp. 62, 77.

Before the reordering, the order is a mix of Urdu and Punjabi
alphabets, with some special characters in the end, and some in
the middle. This is confusing and harder to read and debug.
The reordering makes it more readable.

There are no functional changes. Arabic letters that are unused
in native Saraiki words are left as they are at the end of
the regular expression.

Change-Id: Ibba55d1594c1ffae937412afc6b7ffa8f3ab771a