resourceloader: Remove ES5 fallback for deprecation logger
By removing mw.log.deprecate's dependency on the StringSet polyfill,
we can move most of the mw.log code out of the startup module and
into the regular module 'mediawiki.base' (which, like other modules,
is highly-cached and loaded concurrently).

In this patch, I've limited the logging and StatsD counting to the
ES6-only branch, which means we will no longer collect mw-js-deprecate
stats from IE 11 devices, and IE 11 users will not see console warnings
about use of deprecated MediaWiki JS methods.

The alternative was to collect it but without any deduplication, thus
overrepresenting its users. And to let many console warnings pile up
for the same deprecated method.

This change is in alignment with T178356 and the policy at
<https://www.mediawiki.org/wiki/Compatibility/IE11>.

Change-Id: I68f2021fd39b6ead2f0a3e4110fa41b1932d86ac