resourceloader: Move most of mw.log from startup to mediawiki.base
As of 99d4ad599596c, mw.log.deprecate is no longer used within the
startup module, and I don't expect use of it to return. The only
reason we needed it in startup was because mw.config, which was a hybrid
that is both deprecated (when accessed via window) and not deprecated
(when accessed via mw.config).

For anything else, there is no reason to ever ship deprecated code
in the startup module that isn't internally used by mw.loader.

Even if an mw.loader method were to be deprecated at some point, we'd
remove it from startup and ship it in mediawiki.base instead and
deprecate it there (like '$j' and 'importScript', for example).

For future reference, this move was made possible by preparations in
Ied4d08c554936 and I68f2021fd39b.

Change-Id: I2ef8d4e6c3c32a811d418d48bf2b0da9f3d621b1