Message: Deprecate toString() with an implicit format
This functionality exists to enable the anti-pattern of calling a
format method, without using its return value, and then subsequently
calling toString() to parse the same message, the same way, a second
time with the format you just called.

For example:

```
$msg = new Message('laggedreplicamode');

echo $msg->plain();
// <strong>Warning:</strong> Page may not contain recent
echo $msg->toString();
// <strong>Warning:</strong> Page may not contain recent

echo $msg->escaped();
// &lt;strong&gt;Warning:&lt;/strong&gt; Page may not contain recent
echo $msg->toString();
// &lt;strong&gt;Warning:&lt;/strong&gt; Page may not contain recent
```

Deprecate this rare anti-pattern in favour of using the return value
from the formatting methods such as $msg->plain() and $msg->escaped(),
or for compat we can also support $msg->toString('plain') etc.

This commit also fixes an unintentional lie in the assertion message
of one of the unit tests, which was claiming that toString() is
unaffected by format calls, when in fact it is. (And is what we are
deprecating here).

This implicit formatting and the related $this->format field will
be removed in 1.37.

Bug: T146416
Depends-On: Ia257621fb421854bf13e2965b04009663711ba0a
Change-Id: Ibb22099ff8042406e441d6c7fab87537d3ac92aa
(cherry picked from commit 73e718e3730cea1d94445a646c79265758fcae4c)