jquery: Add method/prop name to jquery_migrate deprecation counter
This is similar to that we did for jQuery 1.9 Migrate in 2014 when
we upgraded from 1.8 to 1.11 (T46740), and this will likely be the
last patch in this series that's suitable for backporting to stable
releases.

After this, we'll start janking stuff out during the 1.37-master
release, slowly working toward its removal.

I've removed the patch file for now since we're far from the original
now, and if any changes do get released by upstream for migrate 3.x
in the next few months we'll have to apply by hand either way.

(Yes, I'm volunteering to do that.)

* Remove needless window.mw guard.
  I think I copied this from our jQuery 1.9 migrate patch,
  when we started the jQuery 3 migration, not realizing that by
  now mw.loader and startup no longer depend on jquery.
  jquery is loaded by an mw.loader callback (as evidenced by the need
  for the window.jQuery lookup pach), so mw.track always exists.
  If it doesn't we have bigger problems.

* Remove migrateTrace code instead of disabling it.

* Change `jquery-migrate` to `jquery_migrate` in mw.track call.
  Otherwise, the name gets transformed into an underscore later on for
  Prometheus, Graphite, Statsv, etc. WMF: in WikimediaEvents/statsd.

* Improve PATCH comments to state upfront whether they change, add,
  or remove something (makes it a little easier for me if/when
  this needs to be rebased).

* Fix duplicate "JQMIGRATE" prefix in cssProps warning.
  Upstream ref https://github.com/jquery/jquery-migrate/pull/424.

Bug: T280944
Change-Id: Ieecf02982468f1c576d2c56b43dcd59d5255bf20