Remove arq as fallback for shy-latn
Having arq as fallback is problematic because shi is LTR and
arq is RTL. In any case, the translation to French
is more complete, so this wouldn't have a significant effect.

It does fix a current problem that wikis with shy as the content
language are shown right to left, which is definitely wrong.

This was requested by User:Reda Kerbouche, and it makes
sense to me.

Bug: T281085
Change-Id: I09769c5ad8512e3545f5de8129e58ab4daec779a