rdbms: Use REQUEST_TIME_FLOAT for LBFactory::getCPInfoFromCookieValue
This means we'll consider a slightly more ChronologyProtector cookies
that are near their expiry. This is perhaps not so useful since that's
also near the time where max replication lag kicks in, and the storing
of the CP data itself is going to be evicted.

But, this may be a more predictable and stable reference point than
"current" time, considering that as of I3140922cec130cad, this
computation now happens lazily at the first call to LBFactory, instead
of during Setup.php, so in a way this change just brings it back to be
closer to the start of the request (except even earlier than it was
before I3140922cec130cad, since before that commit we did wait until
after the first few ~ milliseconds of multiversion, autoloader,
composer, etc.).

Bug: T228895
Change-Id: I76e2130431ab44344eea93ce8f4f8b2322840819