Create buildPersonalPage method for SkinTemplate class, add menu item to personal menu..
Inside buildPersonalUrls method, replace personal_urls variable with method call to build personal page link. Add "user-page" menu to content navigation. This patch is a dependency for a corresponding patch in the Vector repo which needs the new "user-page" menu item to update the personal toolbar based on a feature flag.

Bug: T276561
Change-Id: If4e143aada711d210ae45d33b97a6be0685b6a41