selenium: Fix failing selenium-daily-beta(commons)-MediaWiki Jenkins job
The jobs are running a subset of tests. (Tests containing string
`@daily` in  the name.) To find the tests we are using
`--mochaOpts.grep` WebdriverIO command line option. The filter is
executed after the tests have already started. Since some test files
have no tests with string `@daily`, they are skipped. That results in
JUnit reporter creating a couple of empty XML files. Jenkins thinks
empty files are failures and fails the build.

This commit creates a suite of files that have tests with string
`@daily` and runs only those files.

Bug: T281505
Change-Id: Id844da81feac49f2a2b6820e98ead98d7e932e9f