SECURITY: Fix XSS vulnerability in EditPage and ProtectionForm
The documentation of OutputPage::wrapWikiMsg says it is
equivalent to OutputPage::addWikitextAsInterface called
with wfMessage()->plain(). This is correct, as it parses
it in addWikitextAsInterface. If used with addHtml, no
parsing/escaping is done automatically.

Follow-up to I83f2828bcde160c98a1d97e783a869e64fb4c6ea.

(NOTE: the above, related change set is NOT currently
on a wmf production release branch.  So this issue only
exists on master for now.  This should all be resolved
when both are part of the cut for wmf.4 next week.)

Bug: T281595
Change-Id: Ia6c7a7447e5cea903be9f5d9e76320ea56fedd4c